fix: users-detail 뷰 종속 타입 views로 이동
