+++ apps/web/src/entities/feed/index.ts
@@ -1,4 +1,3 @@
-export { getFeedProfileView, getFeedProfileViewPosts } from './api/client';
 export {
   formatDate,
   formatLikeCount,
@@ -23,7 +22,6 @@ export type {
   FeedPage,
   FeedPost,
   FeedProfilePosts,
-  FeedProfileView,
   PostScope,
   PostSortType,
 } from './model/types';
