+++ apps/web/src/entities/feed/model/types.ts
@@ -2,22 +2,6 @@ import { type PlaceTagValue } from './place-tag';
 
 export type PostScope = 'PUBLIC' | 'PRIVATE';
 
-export type FeedProfileBadge = {
-  id: number;
-  name: string;
-  description: string;
-  imageUrl: string;
-  isAcquired: boolean;
-};
-
-export type FeedProfileMemberInfo = {
-  id?: number;
-  nickname: string;
-  profileImageUrl?: string;
-  introduction: string | null;
-  mainBadge: FeedProfileBadge | null;
-};
-
 export type FeedPost = {
   postId: number;
   name: string;
@@ -45,10 +29,6 @@ export type FeedPage = {
   items: FeedPost[];
   lastPostId: number | null;
   createAt: string | null;
-};
-
-export type FeedProfileView = {
-  memberInfo: FeedProfileMemberInfo;
 };
 
 export type FeedProfilePosts = {
