+++ apps/web/src/views/users-detail/model/types.ts
@@ -1,0 +1,13 @@
+import { type UserBadge } from '@/entities/user';
+
+export type UserDetailProfile = {
+  id?: number;
+  nickname: string;
+  profileImageUrl?: string;
+  introduction: string | null;
+  mainBadge: UserBadge | null;
+};
+
+export type UserDetailData = {
+  memberInfo: UserDetailProfile;
+};
