+++ apps/web/src/views/users-detail/model/use-feed-profile-view-query.ts
@@ -2,14 +2,24 @@
 
 import { useQuery } from '@tanstack/react-query';
 
-import { feedQueryKeys, getFeedProfileView } from '@/entities/feed';
+import { feedQueryKeys } from '@/entities/feed';
+
+import { clientApi } from '@/shared/api/client-api';
+
+import { type UserDetailData } from './types';
+
+function fetchFeedProfileView(memberKey: string) {
+  return clientApi.get<UserDetailData>(
+    `/feed/profileView/${encodeURIComponent(memberKey)}`,
+  );
+}
 
 export function useFeedProfileViewQuery(memberKey: string) {
   const selectedMember = memberKey.trim();
 
   return useQuery({
     queryKey: feedQueryKeys.profileView(selectedMember),
-    queryFn: () => getFeedProfileView(selectedMember),
+    queryFn: () => fetchFeedProfileView(selectedMember),
     enabled: selectedMember.length > 0,
   });
 }
